Add files via upload
This repository consists of two files
1) motion_detector.py
2) Times.csv
The motion_detector.py intakes all the necessary python code to detect the motion using camera. The time in which the motion is recorded in a .csv file, and here the file is Times.csv.
The Times.csv file contains the start time and end time, the start time represents the time in which the object appeared in front of the camera and the end time represents the time in which the object disappears from the front of the camera.